little clean up on player movement code
