+++ UnityProject/Assets/Scripts/Player/Movement/IMovementEffect.cs
@@ -1,0 +1,9 @@
+﻿namespace Player.Movement
+{
+	public interface IMovementEffect
+	{
+		float RunningSpeedModifier { get; }
+		float WalkingSpeedModifier { get; }
+		float CrawlingSpeedModifier { get; }
+	}
+}
