+++ UnityProject/Assets/Scripts/Player/Movement/IMovementEffect.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b138f4c6ea034933ac3a8a88071ec015
+timeCreated: 1631143066
